requirements.txt
streamlit~=1.43.0
numpy~=1.26.4
Pillow~=9.4.0
torch~=2.6.0
torchvision~=0.21.0
scipy~=1.10.1
projet~=0.1.0
setuptools~=58.0.4
requests~=2.32.3
tqdm~=4.65.0
matplotlib~=3.6.3
scikit-learn~=1.6.1